+++ CotEditor/Sources/Settings Window/Other Views/ThemeView.swift
@@ -45,7 +45,7 @@ struct ThemeView: View {
     var body: some View {
         
         HStack(spacing: 0) {
-            ThemeListView(selection: $themeName)
+            ThemeListView(manager: self.manager, selection: $themeName)
             
             Divider()
             
@@ -66,24 +66,20 @@ struct ThemeView: View {
             self.setTheme(name: newValue)
         }
         .task {
-            let changes = NotificationCenter.default
+            let names = NotificationCenter.default
                 .notifications(named: .didUpdateSettingNotification, object: self.manager)
                 .compactMap { $0.userInfo?["change"] as? SettingChange }
+                .compactMap(\.new)
             
-            for await name in changes.compactMap(\.new) {
-                await MainActor.run {
-                    guard
-                        name == self.themeName,
-                        let theme = try? self.manager.setting(name: name)
-                    else { return }
-                    
-                    self.theme = theme
-                }
+            for await name in names where name == self.themeName {
+                self.setTheme(name: name)
             }
         }
         .background()
         .border(.separator)
         .alert(error: $error)
+        .accessibilityElement(children: .contain)
+        .accessibilityLabel(String(localized: "Theme", table: "ThemeEditor"))
     }
     
     
@@ -112,24 +108,308 @@ struct ThemeView: View {
 }
 
 
-private struct ThemeListView: NSViewControllerRepresentable {
+private struct ThemeListView: View {
     
-    typealias NSViewControllerType = ThemeListViewController
+    var manager: ThemeManager
     
     @Binding var selection: String
     
     
-    func makeNSViewController(context: Context) -> ThemeListViewController {
+    @State private var settingNames: [String] = []
+    @State private var exportingItem: TransferableTheme?
+    @State private var deletingItem: String?
+    @FocusState private var editingItem: String?
+    
+    @State private var isExporterPresented = false
+    @State private var isImporterPresented = false
+    @State private var isDeleteConfirmationPresented = false
+    @State private var isImportConfirmationPresented = false
+    @State private var importingError: SettingImportError?
+    @State private var error: (any Error)?
+    
+    
+    var body: some View {
         
-        NSStoryboard(name: "ThemeListView", bundle: nil).instantiateInitialController { coder in
-            ThemeListViewController(coder: coder, selection: $selection)
-        }!
+        VStack(spacing: 0) {
+            List(selection: $selection) {
+                Section(String(localized: "Theme", table: "ThemeEditor")) {
+                    ForEach(self.settingNames, id: \.self) { name in
+                        let state = self.manager.state(of: name)
+                        
+                        SettingNameField(text: name) { newName in
+                            do {
+                                try self.manager.renameSetting(name: name, to: newName)
+                            } catch {
+                                self.error = error
+                                return false
+                            }
+                            self.selection = newName
+                            return true
+                        }
+                        .editDisabled(state?.isBundled == true)
+                        .focused($editingItem, equals: name)
+                        .draggable(TransferableTheme(name: name, canExport: state?.isCustomized == true, data: self.manager.dataForUserSetting(name: name))) {
+                            Label {
+                                Text(name)
+                            } icon: {
+                                Image(nsImage: NSWorkspace.shared.icon(for: .cotTheme))
+                            }
+                        }
+                    }
+                }
+                .listRowSeparator(.hidden)
+            }
+            .dropDestination(for: TransferableTheme.self) { (items, _) in
+                var succeed = false
+                for item in items {
+                    guard let data = item.data() else { continue }
+                    do {
+                        try self.manager.importSetting(data: data, name: item.name, overwrite: false)
+                        succeed = true
+                    } catch let error as SettingImportError {
+                        self.importingError = error
+                        self.isImportConfirmationPresented = true
+                    } catch {
+                        self.error = error
+                    }
+                }
+                return succeed
+            }
+            .contextMenu(forSelectionType: String.self) { selections in
+                if let selection = selections.first {
+                    self.menu(for: selection, isContext: true)
+                }
+            }
+            .listStyle(.bordered)
+            .border(.white)
+            
+            Divider()
+                .padding(.horizontal, 4)
+            
+            HStack {
+                Button {
+                    do {
+                        self.selection = try self.manager.createUntitledSetting()
+                    } catch {
+                        self.error = error
+                    }
+                } label: {
+                    Label(String(localized: "Button.add.label", defaultValue: "Add", table: "Control"), systemImage: "plus")
+                        .padding(2)
+                }
+                .frame(width: 16)
+                
+                Button {
+                    self.deletingItem = self.selection
+                    self.isDeleteConfirmationPresented = true
+                } label: {
+                    Label(String(localized: "Button.remove.label", defaultValue: "Remove", table: "Control"), systemImage: "minus")
+                        .padding(2)
+                }
+                .frame(width: 16)
+                .disabled(self.manager.state(of: self.selection)?.isBundled != false)
+                
+                Spacer()
+                
+                Menu(String(localized: "Button.actions.label", defaultValue: "Actions", comment: "label for action menu button"), systemImage: "ellipsis") {
+                    if #unavailable(macOS 26) {
+                        self.menu(for: self.selection)
+                            .labelStyle(.titleOnly)
+                    } else {
+                        self.menu(for: self.selection)
+                    }
+                }
+                .symbolVariant(.circle)
+            }
+            .labelStyle(.iconOnly)
+            .buttonStyle(.borderless)
+            .padding(6)
+        }
+        .onReceive(self.manager.$settingNames.receive(on: RunLoop.main)) { settingNames in
+            self.settingNames = settingNames
+        }
+        .fileExporter(isPresented: $isExporterPresented, item: self.exportingItem, contentTypes: [.cotTheme], defaultFilename: self.exportingItem?.name) { result in
+            switch result {
+                case .success:
+                    break
+                case .failure(let error):
+                    self.error = error
+            }
+        }
+        .fileImporter(isPresented: $isImporterPresented, allowedContentTypes: [.cotTheme], allowsMultipleSelection: true) { result in
+            switch result {
+                case .success(let urls):
+                    for url in urls {
+                        let accessing = url.startAccessingSecurityScopedResource()
+                        defer {
+                            if accessing {
+                                url.stopAccessingSecurityScopedResource()
+                            }
+                        }
+                        
+                        let name = url.deletingPathExtension().lastPathComponent
+                        do {
+                            let data = try Data(contentsOf: url)
+                            try self.manager.importSetting(data: data, name: name, overwrite: false)
+                        } catch let error as SettingImportError {
+                            self.importingError = error
+                            self.isImportConfirmationPresented = true
+                        } catch {
+                            self.error = error
+                        }
+                    }
+                case .failure(let error):
+                    self.error = error
+            }
+        }
+        .fileDialogConfirmationLabel(String(localized: "Button.import.label", defaultValue: "Import"))
+        .confirmationDialog(String(localized: "ImportDuplicationError.description",
+                                   defaultValue: "“\(self.importingError?.name ?? String(localized: .unknown))” already exists. Do you want to replace it?",
+                                   comment: "%@ is a name of a setting. Refer the same expression by Apple."),
+                            isPresented: $isImportConfirmationPresented, presenting: self.importingError) { item in
+            Button(String(localized: "Button.replace.label", defaultValue: "Replace")) {
+                self.importingError = nil
+                do {
+                    try self.manager.importSetting(data: item.data, name: item.name, overwrite: true)
+                } catch {
+                    self.error = error
+                }
+            }
+            Button("Cancel", role: .cancel) {
+                self.importingError = nil
+            }
+        } message: { _ in
+            Text(String(localized: "ImportDuplicationError.recoverySuggestion",
+                        defaultValue: "A custom setting with the same name already exists. Replacing it will overwrite its current contents.",
+                        comment: "Refer similar expressions by Apple."))
+        }
+        .confirmationDialog(String(localized: "DeletionConfirmationAlert.message",
+                                   defaultValue: "Are you sure you want to delete “\(self.deletingItem ?? String(localized: .unknown))”?"),
+                            isPresented: $isDeleteConfirmationPresented, presenting: self.deletingItem)
+        { name in
+            Button(String(localized: "DeletionConfirmationAlert.button.delete", defaultValue: "Delete"), role: .destructive) {
+                self.deletingItem = nil
+                do {
+                    try self.manager.removeSetting(name: name)
+                } catch {
+                    self.error = error
+                    return
+                }
+                UserDefaults.standard.restore(key: .theme)
+            }
+            Button(String(localized: "Cancel"), role: .cancel) {
+                self.deletingItem = nil
+            }
+        } message: { _ in
+            Text(String(localized: "DeletionConfirmationAlert.informativeText",
+                        defaultValue: "This action cannot be undone."))
+        }
+        .alert(error: $error)
     }
     
     
-    func updateNSViewController(_ nsViewController: ThemeListViewController, context: Context) {
+    /// Builds menu items for either the Action menu button or the context menu.
+    ///
+    /// - Parameters:
+    ///   - selection: The action target.
+    ///   - isContext: Whether the items are for the context menu.
+    /// - Returns: Menu items.
+    @ViewBuilder private func menu(for selection: String, isContext: Bool = false) -> some View {
         
-        nsViewController.select(settingName: self.selection)
+        if let selection = self.manager.state(of: selection) {
+            Button(isContext
+                   ? String(localized: "Action.duplicate.label", defaultValue: "Duplicate")
+                   : String(localized: "Action.duplicate.named.label", defaultValue: "Duplicate “\(selection.name)”"),
+                   systemImage: "plus.square.on.square")
+            {
+                do {
+                    try self.manager.duplicateSetting(name: selection.name)
+                } catch {
+                    self.error = error
+                }
+            }
+            
+            Button(isContext
+                   ? String(localized: "Action.rename.label", defaultValue: "Rename")
+                   : String(localized: "Action.rename.named.label", defaultValue: "Rename “\(selection.name)”"),
+                   systemImage: "pencil")
+            {
+                self.editingItem = selection.name
+            }
+            .disabled(selection.isBundled)
+            
+            Button(isContext
+                   ? String(localized: "Action.restore.label", defaultValue: "Restore")
+                   : String(localized: "Action.restore.named.label", defaultValue: "Restore “\(selection.name)”"),
+                   systemImage: "arrow.clockwise")
+            {
+                do {
+                    try self.manager.restoreSetting(name: selection.name)
+                } catch {
+                    self.error = error
+                }
+            }
+            .disabled(!selection.isRestorable)
+            
+            if isContext {
+                Button(String(localized: "Action.delete.label", defaultValue: "Delete"), systemImage: "trash") {
+                    self.deletingItem = selection.name
+                    self.isDeleteConfirmationPresented = true
+                }
+                .disabled(selection.isBundled)
+            }
+            
+            Button(isContext
+                   ? String(localized: "Action.export.label", defaultValue: "Export…")
+                   : String(localized: "Action.export.named.label", defaultValue: "Export “\(selection.name)”…"),
+                   systemImage: "square.and.arrow.up")
+            {
+                self.exportingItem = TransferableTheme(name: selection.name, data: self.manager.dataForUserSetting(name: selection.name))
+                self.isExporterPresented = true
+            }
+            .modifierKeyAlternate(.option) {
+                Button(isContext
+                       ? String(localized: "Action.revealInFinder.label", defaultValue: "Reveal in Finder")
+                       : String(localized: "Action.revealInFinder.named.label", defaultValue: "Reveal “\(selection.name)” in Finder"),
+                       systemImage: "finder")
+                {
+                    guard let url = self.manager.urlForUserSetting(name: selection.name) else { return }
+                    
+                    NSWorkspace.shared.activateFileViewerSelecting([url])
+                }
+            }
+            .disabled(!selection.isCustomized)
+            
+            if let url = self.manager.urlForUserSetting(name: selection.name) {
+                // -> ShareLink doesn't work for context menu. (macOS 26, 2025-06)
+                if !isContext {
+                    if #available(macOS 26, *) {
+                        ShareLink(item: url)
+                            .disabled(!selection.isCustomized)
+                    } else {
+                        ShareLink(item: url) {
+                            Text(String(localized: "Action.share.label", defaultValue: "Share…"))
+                        }
+                        .disabled(!selection.isCustomized)
+                    }
+                }
+            }
+        }
+        
+        if !isContext {
+            Divider()
+            
+            Button(String(localized: "Action.import.label", defaultValue: "Import…"), systemImage: "square.and.arrow.down") {
+                self.isImporterPresented = true
+            }
+            .modifierKeyAlternate(.option) {
+                Button(String(localized: "Reload All Themes", table: "ThemeEditor"), systemImage: "arrow.clockwise") {
+                    Task.detached(priority: .utility) {
+                        self.manager.loadUserSettings()
+                    }
+                }
+            }
+        }
     }
 }
 
@@ -342,6 +622,39 @@ private extension Theme.SystemDefaultStyle {
         
         get { Color(nsColor: self.color) }
         set { self.color = NSColor(newValue).componentBased }
+    }
+}
+
+
+private struct TransferableTheme: Transferable {
+    
+    var name: String
+    var canExport: Bool
+    var data: @Sendable () -> Data?
+    
+    
+    init(name: String, canExport: Bool = true, data: @autoclosure @Sendable @escaping () -> Data?) {
+        
+        self.name = name
+        self.canExport = canExport
+        self.data = data
+    }
+    
+    
+    static var transferRepresentation: some TransferRepresentation {
+        
+        DataRepresentation(exportedContentType: .cotTheme) { item in
+            guard let data = item.data() else { throw CocoaError(.fileNoSuchFile) }
+            return data
+        }
+        .suggestedFileName { $0.name }
+        .exportingCondition { $0.canExport }
+        
+        FileRepresentation(importedContentType: .cotTheme) { received in
+            let name = received.file.deletingPathExtension().lastPathComponent
+            let data = try Data(contentsOf: received.file)
+            return TransferableTheme(name: name, data: data)
+        }
     }
 }