+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 - The options to highlight matching braces has been removed and they are now always highlighted.
 - Deprecate the feature to change the editor opacity per window.
 - [trivial] Improve the behavior of Add/Remove buttons.
-- [dev] Migrate the syntax list in the Format pane to SwiftUI.
+- [dev] Migrate the theme list and syntax list in the Settings to SwiftUI.
 - [dev] Update the build environment to Xcode 26.0 Beta 3 (Swift 6.2, macOS 26 SDK).
 - [non-AppStore ver.] Add an update badge to the “Check for Update” menu item if an update is available.